refactor(update) improve implementation - Create `defineProperty` util which typesafes around `Object.defineProperty` and gives us better types for constructing update function
